Added copyright
So removing the copyright was the wrong way to handle this. I think just adding on additional notices for future changes is correct. We own the copyright to the changes we made, but Jamie still owns the copyright to the original work. And of course the GPL v3 license is still in effect.